pyproject.toml
[project]
name = "gaij"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "flask",
    "requests",
    "google-api-python-client",
    "google-cloud-firestore",
    "beautifulsoup4",
    "openai",
    "pydantic-settings",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "ruff",
    "black",
    "bandit",
    "radon",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["I", "UP", "B", "C4", "SIM", "N", "RUF"]

[tool.mypy]
strict = true
packages = ["gaij"]
ignore_missing_imports = true

[tool.coverage.run]
source = ["gaij"]
